+++ UnlimitedTowerClient/TowerMain/UnlimitedTower/Assets/Scripts/UIs/TopUIManager.cs
@@ -17,7 +17,7 @@ public class TopUIManager : MonoSingleton<TopUIManager> {
 
     }
 
-    public void ShowBurnResult(List<UserItemData> getItemList, int getUTG)
+    public void ShowBurnResult(List<UserItemData> getItemList, ulong getUTG)
     {
         GameObject objSubView = Instantiate(Resources.Load("UI/Misc/BurnResultVC")) as GameObject;
         objSubView.transform.SetParent(this.transform);
